requirements.txt
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
contextlib2==0.5.4
cycler==0.10.0
decorator==4.0.10
Django==1.8.4
django-autoslug==1.7.2
django-cors-headers==1.1.0
django-crontab==0.7.1
django-extensions==1.5.7
django-geojson==2.8.1
django-leaflet==0.16.0
django-model-utils==2.2
django-taggit==0.17.0
gunicorn==19.3.0
ipython==5.1.0
ipython-genutils==0.1.0
jsonfield==1.0.3
Lemmagen==1.0.3
nltk==3.0.3
numpy==1.10.1
pandas==0.19.2
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.7
psycopg2==2.6.1
ptyprocess==0.5.1
Pygments==2.1.3
pyparsing==2.1.0
python-dateutil==2.4.2
python-memcached==1.58
python-slugify==1.2.1
pytz==2015.7
raven==5.31.0
requests==2.8.1
scikit-learn==0.17.1
scipy==0.16.0
simplegeneric==0.8.1
six==1.10.0
slackclient==1.0.5
traitlets==4.2.2
Unidecode==0.4.19
wcwidth==0.1.7
websocket-client==0.40.0
https://github.com/aljosa/django-tinymce/archive/master.zip
django-autocomplete-light==3.2.10
feedparser==5.2.1
celery==4.1.0
matplotlib==2.1.1